changed logic around getting and posting from firebase
