+++ app/src/main/java/com/example/austin/falldetector/MapsActivity.java
@@ -2,6 +2,7 @@ package com.example.austin.falldetector;
 
 import android.support.v4.app.FragmentActivity;
 import android.os.Bundle;
+import android.util.Log;
 
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -10,11 +11,26 @@ import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
 
 public class MapsActivity extends FragmentActivity implements OnMapReadyCallback {
 
     private GoogleMap mMap;
     fallDetector detector;
+
+    FirebaseDatabase database = FirebaseDatabase.getInstance();
+    DatabaseReference myRef;
+
+    //public static Map<String, String> locations = new HashMap<String,String>();
+    //String timestampA = "";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -25,6 +41,39 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
                 .findFragmentById(R.id.map);
         mapFragment.getMapAsync(this);
+
+        // Read from the database
+        myRef = database.getReference("falls");
+        myRef.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                // This method is called once with the initial value and again
+                // whenever data at this location is updated.
+
+                for (DataSnapshot snapshotChild: dataSnapshot.getChildren()) {
+                    String value = snapshotChild.getValue(String.class);
+                    if( value != null ){
+                        Log.d("Get Data", value);
+                    }
+                }
+                Log.d("Firebase Updated", "yay");
+
+                /*
+                String value = dataSnapshot.getValue(String.class);
+                if(value != null){
+                    Log.d("Firebase Updated", value);
+                    //locations.put(Long.toString(System.currentTimeMillis()), value);
+                    //timestampA = Long.toString(System.currentTimeMillis());
+                }
+                */
+            }
+
+            @Override
+            public void onCancelled(DatabaseError error) {
+                // Failed to read value
+                Log.w("FireBaseError", "Failed to read value.", error.toException());
+            }
+        });
     }
